+++ data-raw/yocotopuce/plot-q36-data-pkg.R
@@ -9,7 +9,7 @@ library(AS7343)
 # times are out of sync!!
 
 Q36_01.tb <-
-  read_yocto_spectral_csv(file = "data-raw/yocotopuce/spctrl-01-Lumitronix-Q36-150mm.csv",
+  read_yocto_spctlog(file = "data-raw/yocotopuce/spctrl-01-Lumitronix-Q36-150mm.csv",
                           tz = "Europe/Helsinki")
 Q36_01.tb <- subset(Q36_01.tb, select = !grepl("min|max", names(Q36_01.tb))) |>
   subset(month(time) == 10)
@@ -17,7 +17,7 @@ Q36_01.tb <- subset(Q36_01.tb, select = !grepl("min|max", names(Q36_01.tb))) |>
 time.s01.at.max <- with(Q36_01.tb, time[VIS.avg == (max(VIS.avg))])
 
 Q36_02.tb <-
-  read_yocto_spectral_csv(file = "data-raw/yocotopuce/spctrl-02-Lumitronix-Q36-150mm.csv",
+  read_yocto_spctlog(file = "data-raw/yocotopuce/spctrl-02-Lumitronix-Q36-150mm.csv",
                           tz = "Europe/Helsinki")
 Q36_02.tb <- subset(Q36_02.tb, select = !grepl("min|max", names(Q36_02.tb))) |>
   subset(month(time) == 10)